requirements.txt
aiogram==3.4.1
python-dotenv==1.0.1
aiohttp==3.9.5
openai==1.30.1